+++ Marlin/src/types.h
@@ -23,6 +23,8 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
-typedef unsigned long millis_t;
+#include <stdint.h>
+
+typedef uint32_t millis_t;
 
 #endif
